Fixed some errors and implement user authorization for Element Actions in File Explorer

+++ EiDiA_backend/src/controllers/record.js
@@ -372,27 +372,34 @@ const deleteDocument = (req, res) => {
     }
     DocumentModel.findById({_id: req.params.documentId})
         .then((document) => {
-                DocumentModel.deleteOne({_id: req.params.documentId})
-                    .then(result => {
+                if (document.createdBy === req.userId || req.userRole === 'admin') {
+                    DocumentModel.deleteOne({_id: req.params.documentId})
+                        .then(result => {
 
-                            res.status(200).json(result);
-                            LogModel.create({
-                                userId: req.userId,
-                                recordId: document.recordId,
-                                log: "deleted " + document.fileType + ": " + document.name
-                            }).then("Created Log").catch((e) => {
-                                console.error("Didnt Create Log" + e)
+                                res.status(200).json(result);
+                                LogModel.create({
+                                    userId: req.userId,
+                                    recordId: document.recordId,
+                                    log: "deleted " + document.fileType + ": " + document.name
+                                }).then("Created Log").catch((e) => {
+                                    console.error("Didnt Create Log" + e)
+                                })
+                            }
+                        )
+                        .catch((error) => {
+                            console.log(error.message)
+                            res.status(500).json({
+
+                                error: "Internal Server error",
+                                message: error.message,
                             })
-                        }
-                    )
-                    .catch((error) => {
-                        console.log(error.message)
-                        res.status(500).json({
-
-                            error: "Internal Server error",
-                            message: error.message,
                         })
+                } else {
+                    return res.status(400).json({
+                        error: "Not authorised to make this delete Request"
                     })
+                }
+
             }
         )
         .catch((error) => {
